requirements.txt
amqp==5.3.1
asgiref==3.8.1
babel==2.17.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
Django==5.1.7
django-fsm==3.0.0
django-money==3.5.3
djangorestframework==3.15.2
drf-yasg==1.21.10
Genshi==0.7.9
idna==3.10
inflection==0.5.1
kombu==5.5.0
lxml==5.3.1
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pillow==11.1.0
polib==1.2.0
prompt_toolkit==3.0.50
py-moneyed==3.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-sql==1.5.2
python-stdnum==1.20
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
relatorio==0.11.1
reportlab==4.3.1
requests==2.32.3
setuptools==77.0.1
simpleeval==1.0.3
six==1.17.0
sqlparse==0.5.3
trytond==7.4.8
trytond_account==7.4.5
trytond_account_invoice==7.4.2
trytond_account_product==7.4.0
trytond_company==7.4.1
trytond_country==7.4.0
trytond_currency==7.4.0
trytond_party==7.4.1
trytond_product==7.4.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3

python-dotenv~=1.0.1